html/semantics/embedded-content-0/media-elements/error_property_exists.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.error - existence test</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-error">spec reference</a></p>
  <audio id="a">
  </audio>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_idl_attribute(
   document.getElementById("a"),
   "error",
   "audioElement.error property exists");
}, "audio.error - existence test");

test(function() {
 assert_idl_attribute(
   document.getElementById("v"),
   "error",
   "videoElement.error property exists");
}, "video.error - existence test");
  </script>
 </body>
</html>
